gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "aio-pika>=9.5.5",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
